Added missing docstrings that were preventing Javadoc generation.
(Will need to fix up some of them better, but it's good enough for now).

Modified .gitignore to allow the commital of said Javadocs.

Moved leftover classes from the original lab version to a temporary storage folder in root to keep it out of the way of the compiler.
(Will extract anything useful from it later, then delete the rest.)